refactor(view): move banner/context/logo utilities into view_banner.go (move-only)
