+++ cmd/app/view.go
@@ -1,13 +1,12 @@
 package main
 
 import (
-	"fmt"
-	"image/color"
-	"net/url"
-	"os"
-	"regexp"
-	"strings"
-	"time"
+    "fmt"
+    "image/color"
+    "os"
+    "regexp"
+    "strings"
+    "time"
 
 	"github.com/charmbracelet/lipgloss/v2"
 	"github.com/darksworm/argonaut/pkg/model"
@@ -236,179 +235,27 @@ func countLines(s string) int {
 }
 
 // renderBanner - 1:1 mapping from Banner.tsx
-func (m Model) renderBanner() string {
-	// Determine narrow layout threshold similar to TS
-	isNarrow := m.state.Terminal.Cols <= 100
-
-	// Cyan badge for very narrow terminals
-	if isNarrow {
-		badge := m.renderSmallBadge(false)
-
-		// Add spacing before and after the badge, and after Project line
-		var sections []string
-		sections = append(sections, "") // Empty line before badge
-		sections = append(sections, badge)
-		sections = append(sections, "") // Empty line after badge
-
-		// Context block (stacked)
-		ctx := m.renderContextBlock(true)
-		sections = append(sections, ctx)
-		sections = append(sections, "") // Empty line after Project
-
-		return strings.Join(sections, "\n")
-	}
-
-	// Wide layout: left context block, right ASCII logo, bottom-aligned and pushed to right edge
-	left := m.renderContextBlock(false)
-	right := m.renderAsciiLogo()
-
-	// Normalize heights by padding top of the shorter block
-	leftLines := strings.Count(left, "\n") + 1
-	rightLines := strings.Count(right, "\n") + 1
-	if leftLines < rightLines {
-		pad := strings.Repeat("\n", rightLines-leftLines)
-		left = pad + left
-	} else if rightLines < leftLines {
-		pad := strings.Repeat("\n", leftLines-rightLines)
-		right = pad + right
-	}
-
-	// Compute full row width inside main container (account for main container padding of 1 on each side)
-	total := max(0, m.state.Terminal.Cols-2)
-	return joinWithRightAlignment(left, right, total)
-}
+// moved to view_banner.go
 
 // renderSmallBadge renders the compact "Argonaut <version>" badge used in narrow terminals.
 // If grayscale is true, it uses a gray background with dark text so it stays readable when
 // the base layer is desaturated for modal backdrops.
-func (m Model) renderSmallBadge(grayscale bool) string {
-	st := lipgloss.NewStyle().
-		Bold(true).
-		PaddingLeft(1).
-		PaddingRight(1)
-
-	if grayscale {
-		// Medium gray background with near-black foreground for contrast
-		st = st.Background(lipgloss.Color("243")).Foreground(lipgloss.Color("16"))
-	} else {
-		st = st.Background(cyanBright).Foreground(whiteBright)
-	}
-	return st.Render("Argonaut " + appVersion)
-}
+// moved to view_banner.go
 
 // renderContextBlock renders the left-side context (labels + values)
-func (m Model) renderContextBlock(isNarrow bool) string {
-	if m.state.Server == nil {
-		return ""
-	}
-
-	label := lipgloss.NewStyle().Bold(true).Foreground(whiteBright)
-	cyan := lipgloss.NewStyle().Foreground(cyanBright)
-	green := lipgloss.NewStyle().Foreground(syncedColor)
-
-	// Values
-	serverHost := hostFromURL(m.state.Server.BaseURL)
-	clusterScope := scopeToText(m.state.Selections.ScopeClusters)
-	namespaceScope := scopeToText(m.state.Selections.ScopeNamespaces)
-	projectScope := scopeToText(m.state.Selections.ScopeProjects)
-
-	var lines []string
-	lines = append(lines, fmt.Sprintf("%s %s", label.Render("Context:"), cyan.Render(serverHost)))
-	if clusterScope != "—" {
-		lines = append(lines, fmt.Sprintf("%s %s", label.Render("Cluster:"), clusterScope))
-	}
-	if namespaceScope != "—" {
-		lines = append(lines, fmt.Sprintf("%s %s", label.Render("Namespace:"), namespaceScope))
-	}
-	if projectScope != "—" {
-		lines = append(lines, fmt.Sprintf("%s %s", label.Render("Project:"), projectScope))
-	}
-	if !isNarrow && m.state.APIVersion != "" {
-		lines = append(lines, fmt.Sprintf("%s %s", label.Render("ArgoCD:"), green.Render(m.state.APIVersion)))
-	}
-
-	// Right padding between context and logo
-	block := strings.Join(lines, "\n")
-	return lipgloss.NewStyle().PaddingRight(2).Render(block)
-}
+// moved to view_banner.go
 
 // renderAsciiLogo renders the right-side Argonaut ASCII logo like TS component
-func (m Model) renderAsciiLogo() string {
-	cyan := lipgloss.NewStyle().Foreground(cyanBright)
-	white := lipgloss.NewStyle().Foreground(whiteBright)
-	dim := lipgloss.NewStyle().Foreground(dimColor)
-
-	// Last line version (Argonaut version from build)
-	version := appVersion
-	versionPadded := fmt.Sprintf("%13s", version)
-
-	l1 := cyan.Render("   _____") + strings.Repeat(" ", 43) + white.Render(" __   ")
-	l2 := cyan.Render("  /  _  \\_______  ____   ____") + white.Render("   ____ _____   __ ___/  |_ ")
-	l3 := cyan.Render(" /  /_\\  \\_  __ \\/ ___\\ /  _ \\ ") + white.Render("/    \\\\__  \\ |  |  \\   __\\")
-	l4 := cyan.Render(" /    |    \\  | \\/ /_/  >  <_> )  ") + white.Render(" |  \\/ __ \\|  |  /|  |  ")
-	l5 := cyan.Render("\\____|__  /__|  \\___  / \\____/") + white.Render("|___|  (____  /____/ |__|  ")
-	l6 := cyan.Render("        \\/     /_____/             ") + white.Render("\\/     \\/") + dim.Render(versionPadded)
-
-	return strings.Join([]string{l1, l2, l3, l4, l5, l6}, "\n")
-}
+// moved to view_banner.go
 
 // scopeToText formats a selection set for display
-func scopeToText(set map[string]bool) string {
-	if len(set) == 0 {
-		return "—"
-	}
-	vals := make([]string, 0, len(set))
-	for k := range set {
-		vals = append(vals, k)
-	}
-	sortStrings(vals)
-	return strings.Join(vals, ",")
-}
+// moved to view_banner.go
 
 // hostFromURL extracts host from URL (similar to TS hostFromUrl)
-func hostFromURL(s string) string {
-	if s == "" {
-		return "—"
-	}
-	if u, err := url.Parse(s); err == nil && u.Host != "" {
-		return u.Host
-	}
-	return s
-}
+// moved to view_banner.go
 
 // joinWithRightAlignment composes two multi-line blocks with the right block flush to the given width
-func joinWithRightAlignment(left, right string, totalWidth int) string {
-	leftLines := strings.Split(left, "\n")
-	rightLines := strings.Split(right, "\n")
-	n := len(leftLines)
-	if len(rightLines) > n {
-		n = len(rightLines)
-	}
-
-	// Ensure equal length slices
-	if len(leftLines) < n {
-		pad := make([]string, n-len(leftLines))
-		leftLines = append(pad, leftLines...)
-	}
-	if len(rightLines) < n {
-		pad := make([]string, n-len(rightLines))
-		rightLines = append(pad, rightLines...)
-	}
-
-	var out []string
-	for i := 0; i < n; i++ {
-		l := leftLines[i]
-		r := rightLines[i]
-		lw := lipgloss.Width(l)
-		rw := lipgloss.Width(r)
-		filler := totalWidth - lw - rw
-		if filler < 1 {
-			filler = 1
-		}
-		out = append(out, l+strings.Repeat(" ", filler)+r)
-	}
-	return strings.Join(out, "\n")
-}
+// moved to view_banner.go
 
 // contentInnerWidth computes inner content width inside the bordered box
 func (m Model) contentInnerWidth() int {
